more replacements of TransitionTrie with VerkleTrie
1. TransitionTrie doesn't work at this stage, we are only testing the
binary trees at genesis, so let's make sure TransitionTrie isn't
activated anywhere.
2. There was a superfluous check for Transitioned(), which actually made
the code take the wrong turn: if the verkle fork has occured, and if the
transition isn't in progress, it means that the conversion has ended*

* to be complete, it could be that this is in the period before the
start of the iterator sweep, but in that case the transition tree also
needs to be returned. So there is a missing granularity in this code,
but it's ok at this stage since the transition isn't supported yet.